Refactor behave environment definition
* environment.py now just contains the behave testing hooks
* helper definitions moved to `leapp_testing` support library
